Add files via upload
design_1 is the mobile PHP design for the daycare sponsor I found for this project.  It has the daycare's logo and has the link to this repository.  It gives options to login with Facebook, twitter, or to register using a form.  This allows the daycare to collect information from users who login and allows to target them with the daycare's Facebook or Twitter account as well.

splash_template is the basic template